+++ cmd/skywire-cli/commands/visor/ping.go
@@ -12,16 +12,20 @@ import (
 	"github.com/skycoin/skywire/pkg/visor"
 )
 
-var tries int
-var pcktSize int
+var (
+	tries       int
+	pcktSize    int
+	pubVisCount int
+)
 
 func init() {
 	RootCmd.AddCommand(pingCmd)
-	pingCmd.Flags().IntVarP(&tries, "tries", "t", 1, "Number of pings")
-	pingCmd.Flags().IntVarP(&pcktSize, "size", "s", 32, "Size of packet, in KB, default is 32KB")
+	pingCmd.Flags().IntVarP(&tries, "tries", "t", 1, "Number of tries")
+	pingCmd.Flags().IntVarP(&pcktSize, "size", "s", 2, "Size of packet, in KB, default is 2KB")
 	RootCmd.AddCommand(testCmd)
-	testCmd.Flags().IntVarP(&tries, "tries", "t", 1, "Number of tests per public visors")
-	testCmd.Flags().IntVarP(&pcktSize, "size", "s", 32, "Size of packet, in KB, default is 32KB")
+	testCmd.Flags().IntVarP(&tries, "tries", "t", 1, "Number of tries per public visors")
+	testCmd.Flags().IntVarP(&pcktSize, "size", "s", 2, "Size of packet, in KB, default is 2KB")
+	testCmd.Flags().IntVarP(&pubVisCount, "count", "c", 2, "Count of Public Visors for using in test.")
 }
 
 var pingCmd = &cobra.Command{
@@ -55,7 +59,7 @@ var testCmd = &cobra.Command{
 	Short: "Test the visor with public visors on network",
 	Long:  "\n	Creates a route with public visors as a hop and returns latency on the conn",
 	Run: func(cmd *cobra.Command, args []string) {
-		pingConfig := visor.PingConfig{Tries: tries, PcktSize: pcktSize}
+		pingConfig := visor.PingConfig{Tries: tries, PcktSize: pcktSize, PubVisCount: pubVisCount}
 		rpcClient, err := clirpc.Client(cmd.Flags())
 		if err != nil {
 			os.Exit(1)
@@ -63,7 +67,7 @@ var testCmd = &cobra.Command{
 		results, err := rpcClient.TestVisor(pingConfig)
 		internal.Catch(cmd.Flags(), err)
 		for i, result := range results {
-			internal.PrintOutput(cmd.Flags(), result, fmt.Sprintf("Test No. %d\nPK: %s\nMax: %s\nMin: %s\nMean: %s\n\n", i+1, result.PK, result.Max, result.Min, result.Mean))
+			internal.PrintOutput(cmd.Flags(), result, fmt.Sprintf("Test No. %d\nPK: %s\nMax: %s\nMin: %s\nMean: %s\nStatus: %s\n\n", i+1, result.PK, result.Max, result.Min, result.Mean, result.Status))
 		}
 	},
 }